# api/openapi-spec/v1_swagger.yaml
basePath: /api/v1
definitions:
  request.Authentication:
    properties:
      hash:
        description: |-
          The player's Mojang session hash.
          example: 069a79f444e94726a5befca90e38aaf5
        type: string
      token:
        description: |-
          The player's Discord token.
          example: OTk1MTU1NzcyMzYxMTQ2NDM4.AAAAAA.BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB
        type: string
      username:
        description: |-
          The player's username.
          example: Notch
        type: string
    required:
    - hash
    - token
    - username
    type: object
  request.Settings:
    properties:
      listed:
        description: |-
          Whether the party can be listed or not.
          example: true
        type: boolean
      max_players:
        description: |-
          The maximum number of players in the party.
          example: 10
        type: integer
      public:
        description: |-
          Whether the party is public or not.
          If false can only be joined by invite.
          example: true
        type: boolean
    required:
    - listed
    - max_players
    - public
    type: object
  response.Authentication:
    properties:
      access_token:
        description: |-
          The access token to use for the API
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
      expires_in:
        description: |-
          The duration of the token (in seconds).
          example: 3600
        type: integer
      token_type:
        description: |-
          The type of the token.
          example: Bearer
        type: string
    type: object
  response.Party:
    properties:
      creation:
        description: |-
          The creation date of the party.
          example: 2021-10-10T12:00:00Z
        type: string
      id:
        description: |-
          The ID of the party.
          It is a random string of 30 characters.
        type: string
      join_secret:
        description: |-
          The join secret of the party.
          It is a random string of 100 characters.
        type: string
      leader:
        allOf:
        - $ref: '#/definitions/response.Player'
        description: The leader of the party.
      players:
        description: The list of players in the party.
        items:
          $ref: '#/definitions/response.Player'
        type: array
      settings:
        allOf:
        - $ref: '#/definitions/response.Settings'
        description: The settings of the party.
    type: object
  response.Player:
    properties:
      discord_id:
        description: |-
          The player's Discord ID.
          example: "385441179069579265"
        type: string
      id:
        description: |-
          The player's UUID.
          example: 069a79f4-44e9-4726-a5be-fca90e38aaf5
        type: string
      name:
        description: |-
          The player's name.
          example: Notch
        type: string
    type: object
  response.Settings:
    properties:
      listed:
        description: |-
          Whether the party can be listed or not.
          example: true
        type: boolean
      max_players:
        description: |-
          The maximum number of players in the party.
          example: 10
        type: integer
      public:
        description: |-
          Whether the party is public or not.
          If false can only be joined by invite.
          example: true
        type: boolean
    required:
    - listed
    - max_players
    - public
    type: object
info:
  contact:
    name: Lambda Discord
    url: https://discord.gg/J23U4YEaAr
  description: This is the API for the Lambda Discord RPC handler
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
  title: Lambda RPC API
  version: "1.0"
paths:
  /party/create:
    post:
      consumes:
      - application/json
      description: Create a new party
      parameters:
      - description: Settings
        in: body
        name: Settings
        schema:
          $ref: '#/definitions/request.Settings'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Party'
      security:
      - Bearer: []
      summary: Create a new party
      tags:
      - Party
  /party/join:
    put:
      consumes:
      - application/json
      description: Join a party
      parameters:
      - description: Party ID
        in: body
        name: ID
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/response.Party'
      security:
      - Bearer: []
      summary: Join a party
      tags:
      - Party
  /party/login:
    post:
      consumes:
      - application/json
      description: Login to the server using a Discord identify token, a Minecraft
        username and a Mojang session hash
      parameters:
      - description: Authentication
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/request.Authentication'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Authentication'
      summary: Login to the server
      tags:
      - Party
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"